requirements.txt
absl-py==1.4.0
addict==2.4.0
# -e git+git@github.com:ALRhub/air_hockey.git@3bf068b7fa72dcd9c713b9fdc971cfdfd711d272#egg=AirHockeyChallenge&subdirectory=../../air_hockey
ale-py==0.7.4
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
esdk-obs-python==3.22.2
exceptiongroup==1.1.1
# -e git+https://github.com/Thifhi/fancy_gym_dev@3ca23b36fdea21fbddb0dc0c407eefbb475c2eaf#egg=fancy_gym&subdirectory=../../fancy_gym_dev
filelock==3.12.0
fonttools==4.39.3
glfw==2.5.9
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
gym==0.25.2
gym-notices==0.0.8
idna==3.4
imageio==2.27.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mp-pytorch==0.1.3
mpmath==1.3.0
mujoco==2.3.4
mushroom-rl==1.9.1
networkx==3.1
nlopt==2.7.1
numpy==1.23.5
numpy-ml==0.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
osqp==0.6.2.post9
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pluggy==1.0.0
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.10.1
pygame==2.3.0
Pygments==2.15.1
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
qdldl==0.1.7
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stable-baselines3
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
# torch==2.0.0+cpu
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
